refactored dataset calculations into separate classes
